requirements.txt
torch
torchvision
numpy
matplotlib
PyYAML
